# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Migration Factory API",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create Project",
    "operationId": "create_project_api_v1_projects_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List Projects",
    "operationId": "list_projects_api_v1_projects_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ProjectOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/sources/browse": {
   "get": {
    "tags": [
     "sources"
    ],
    "summary": "Browse",
    "description": "Folders as the worker sees them, only inside MF_LOCAL_SOURCE_ROOTS. Directory names only (no files, no\ncontents); symlinked and hidden folders are not listed; '..' or links leaving a root are rejected.",
    "operationId": "browse_api_v1_sources_browse_get",
    "parameters": [
     {
      "name": "path",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 1000
        },
        {
         "type": "null"
        }
       ],
       "title": "Path"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BrowseOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/uploads": {
   "post": {
    "tags": [
     "sources"
    ],
    "summary": "Upload Zip",
    "description": "Body = the ZIP bytes (Content-Type: application/zip). Size enforced while streaming, structure validated before storing.",
    "operationId": "upload_zip_api_v1_uploads_post",
    "parameters": [
     {
      "name": "x-filename",
      "in": "header",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 255,
       "default": "project.zip",
       "title": "X-Filename"
      }
     },
     {
      "name": "content-length",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Content-Length"
      }
     }
    ],
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UploadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/scans/{scan_id}/workspace-migrations": {
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Start",
    "operationId": "start_api_v1_scans__scan_id__workspace_migrations_post",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WorkspaceMigrationIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "List For Scan",
    "operationId": "list_for_scan_api_v1_scans__scan_id__workspace_migrations_get",
    "parameters": [
     {
      "name": "scan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/WorkspaceMigrationOut"
         },
         "title": "Response List For Scan Api V1 Scans  Scan Id  Workspace Migrations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/workspace-migrations/{wm_id}": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "Get One",
    "operationId": "get_one_api_v1_workspace_migrations__wm_id__get",
    "parameters": [
     {
      "name": "wm_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Wm Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WorkspaceMigrationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/workspace-migrations/{wm_id}/retry": {
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Retry",
    "description": "Doc 20 \u00a720: retry some projects (or resume every non-successful one); successful projects are reused, not repeated,\nand dependents blocked only by the retried projects become READY.",
    "operationId": "retry_api_v1_workspace_migrations__wm_id__retry_post",
    "parameters": [
     {
      "name": "wm_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Wm Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WorkspaceRetryIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "BrowseEntry": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "path": {
      "type": "string",
      "title": "Path"
     },
     "hasPom": {
      "type": "boolean",
      "title": "Haspom"
     }
    },
    "type": "object",
    "required": [
     "name",
     "path",
     "hasPom"
    ],
    "title": "BrowseEntry"
   },
   "BrowseOut": {
    "properties": {
     "path": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Path"
     },
     "parent": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent"
     },
     "hasPom": {
      "type": "boolean",
      "title": "Haspom"
     },
     "entries": {
      "items": {
       "$ref": "#/components/schemas/BrowseEntry"
      },
      "type": "array",
      "title": "Entries"
     },
     "truncated": {
      "type": "boolean",
      "title": "Truncated",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "path",
     "parent",
     "hasPom",
     "entries"
    ],
    "title": "BrowseOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "JobOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "phase": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phase"
     },
     "progress": {
      "type": "integer",
      "title": "Progress"
     },
     "attempts": {
      "type": "integer",
      "title": "Attempts"
     },
     "cancelRequested": {
      "type": "boolean",
      "title": "Cancelrequested"
     },
     "errorCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errorcode"
     },
     "errorMessage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errormessage"
     },
     "createdBy": {
      "type": "string",
      "title": "Createdby"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "startedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Startedat"
     },
     "finishedAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finishedat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "state",
     "phase",
     "progress",
     "attempts",
     "cancelRequested",
     "errorCode",
     "errorMessage",
     "createdBy",
     "createdAt",
     "startedAt",
     "finishedAt"
    ],
    "title": "JobOut"
   },
...
   "Page_ProjectOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ProjectOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "nextCursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nextcursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[ProjectOut]"
   },
...
   "ProjectCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "pattern": "^[\\w .\\-()]+$",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "maxLength": 2000,
      "title": "Description",
      "default": ""
     },
     "sourceType": {
      "type": "string",
      "enum": [
       "local",
       "zip",
       "git"
      ],
      "title": "Sourcetype"
     },
     "targetRuntime": {
      "type": "string",
      "enum": [
       "spring",
       "quarkus"
      ],
      "title": "Targetruntime"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "defaultBranch": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[\\w./-]{1,200}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Defaultbranch"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "uploadId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Uploadid"
     }
    },
    "type": "object",
    "required": [
     "name",
     "sourceType",
     "targetRuntime"
    ],
    "title": "ProjectCreate"
   },
   "ProjectOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "sourceType": {
      "type": "string",
      "title": "Sourcetype"
     },
     "repositoryUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repositoryurl"
     },
     "repositoryProvider": {
      "type": "string",
      "title": "Repositoryprovider"
     },
     "credentialRef": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Credentialref"
     },
     "localPath": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Localpath"
     },
     "upload": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Upload"
     },
     "defaultBranch": {
      "type": "string",
      "title": "Defaultbranch"
     },
     "targetRuntime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Targetruntime"
     },
     "requiredGates": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Requiredgates"
     },
     "independentReview": {
      "type": "boolean",
      "title": "Independentreview"
     },
     "prConfig": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prconfig"
     },
     "myRole": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Myrole"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "lastScan": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lastscan"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "description",
     "sourceType",
     "repositoryUrl",
     "repositoryProvider",
     "credentialRef",
     "defaultBranch",
     "targetRuntime",
     "requiredGates",
     "independentReview",
     "prConfig",
     "myRole",
     "version",
     "createdAt"
    ],
    "title": "ProjectOut"
   },
...
   "UploadOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "filename": {
      "type": "string",
      "title": "Filename"
     },
     "sha256": {
      "type": "string",
      "title": "Sha256"
     },
     "sizeBytes": {
      "type": "integer",
      "title": "Sizebytes"
     },
     "structure": {
      "additionalProperties": true,
      "type": "object",
      "title": "Structure"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     }
    },
    "type": "object",
    "required": [
     "id",
     "filename",
     "sha256",
     "sizeBytes",
     "structure",
     "createdAt"
    ],
    "title": "UploadOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WorkspaceMigrationIn": {
    "properties": {
     "profileId": {
      "type": "string",
      "format": "uuid",
      "title": "Profileid"
     },
     "java": {
      "type": "string",
      "enum": [
       "17",
       "21"
      ],
      "title": "Java",
      "default": "21"
     },
     "accept": {
      "patternProperties": {
       "^[A-Z0-9_]{2,60}$": {
        "type": "string",
        "maxLength": 500,
        "minLength": 10
       }
      },
      "type": "object",
      "title": "Accept"
     },
     "approve": {
      "items": {
       "type": "string",
       "pattern": "^[a-z0-9-]{2,60}$"
      },
      "type": "array",
      "maxItems": 30,
      "title": "Approve"
     }
    },
    "type": "object",
    "required": [
     "profileId"
    ],
    "title": "WorkspaceMigrationIn"
   },
   "WorkspaceMigrationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "retryOf": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retryof"
     },
     "retry": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Retry",
      "default": []
     },
     "projectId": {
      "type": "string",
      "format": "uuid",
      "title": "Projectid"
     },
     "scanId": {
      "type": "string",
      "format": "uuid",
      "title": "Scanid"
     },
     "profileId": {
      "type": "string",
      "format": "uuid",
      "title": "Profileid"
     },
     "jobId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Jobid"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "java": {
      "type": "string",
      "title": "Java"
     },
     "accepted": {
      "additionalProperties": true,
      "type": "object",
      "title": "Accepted"
     },
     "approved": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Approved"
     },
     "results": {
      "additionalProperties": true,
      "type": "object",
      "title": "Results"
     },
     "summary": {
      "additionalProperties": true,
      "type": "object",
      "title": "Summary"
     },
     "createdBy": {
      "type": "string",
      "title": "Createdby"
     },
     "createdAt": {
      "type": "string",
      "format": "date-time",
      "title": "Createdat"
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "projectId",
     "scanId",
     "profileId",
     "jobId",
     "state",
     "java",
     "accepted",
     "approved",
     "results",
     "summary",
     "createdBy",
     "createdAt"
    ],
    "title": "WorkspaceMigrationOut"
   },
   "WorkspaceRetryIn": {
    "properties": {
     "projects": {
      "items": {
       "type": "string",
       "pattern": "^[\\w.\\-/]{1,200}$"
      },
      "type": "array",
      "maxItems": 100,
      "title": "Projects"
     }
    },
    "type": "object",
    "title": "WorkspaceRetryIn"
   }
  }
 }
}